Evaluation

Main evaluation function

cereal.ceval.main — Function

Evaluation function

main( locator::Function , N::Number , q::Real , k::Number ,     
      counter::Bool , usexgen::Bool )

This function tests the user specified locator function for N stochastically generated test cases. The results produced by the locator are compared (by way of the comp function) to the generated intersection points up to a threshold value of q. The variable k is the number of emission points to generate for each test case, and the variable usexgen replaces the function pgen with xgen for test case generation.

Examples:

cereal.ceval.main(cereal.locatorfunc(4,"CFM10"),1e5,1e-6,4)

cereal.ceval.main(cereal.locatorfunc(4,"FHC21"),1e5,1e-6,4)

cereal.ceval.main(cereal.locatorfunc(5,"RTC21"),1e5,1e-9,5)

cereal.ceval.main(cereal.locatorfunc(6,"RTC21"),1e5,5e-13,6)

Test case generator functions

cereal.ceval.vgenerator — Function

Vector generator

vgenerator( tpfl::DataType )

This function generates a random 3-vector of unit length. Returns a three component vector.

cereal.ceval.nullgen — Function

Null vector generator

nullgen( tpfl::DataType )

This function generates a random past directed null vector. Returns a four component vector.

cereal.ceval.pgen — Function

Intersection and emission point generator

pgen( tpfl::DataType , N::Int )

This function generates a point Xc and N random emission points (in a $4×$N matrix X) on the past null cone of Xc. Returns a tuple (X,Xc).

cereal.ceval.xgen — Function

Restricted intersection and emission point generator

xgen( xc::Real , r1::Real , r2::Real , N::Int )

This function generates a point Xc=[0;xc;0;0] and N random emission points (in a $4×$N matrix X) on the past null cone of Xc at a radius r such that r1<r<r2. Returns a tuple (X,Xc).

Test examples