Locator Functions

Locator function selector

cereal.locatorfunc — Function

Locator function

locatorfunc( N::Int , Method::String )

This function selects the locator function to use. The first argument N determines the minimum number of emission points that the locator function will use. N=4 or N≥5 is recommended. The second argument Method selects the formula or algorithm used to compute the intersection point $X_c$. The available methods are "FHC21", "CFM10", and for N≥5, "RTC21".

By default, the locator function assumes N=5 and Method="RTC21":

julia> locatorfunc() == locatorfunc(5,"RTC21")
true

Basic method functions

cereal.locator5RTC21 — Function

Five point locator function (RTC21)

locator5RTC21( X::RealMtx )

The function presented here implements the five point relativistic location formula given by Ruggiero, Tartaglia, and Casalino in Ruggiero et al., arxiv:2111.13423 (2021).

cereal.locator4CFM10 — Function

Four point locator function (CFM10)

locator4CFM10( X::RealMtx )

The function presented here implements the four point relativistic location formula given by Coll, Ferrando, and Morales-Lladosa:

$X_c = X_4 + y_* - \frac{η(y_*,y_*)}{η(y_*,χ)±√Δ} χ$.

See Coll et al., Class.Quant.Grav. 27 (2010) 065013 and Coll et al., Phys. Rev. D 86, 084036 (2012) for details.

cereal.locator4FHC21 — Function

Four point locator function (FHC21)

locator4FHC21( X::RealMtx )

This function implements the four point relativistic location algorithm of the authors based on performing transformations to adapted frames.

The algorithm is conceptually simple for a spacelike hyperplane:

Find and perform a Lorentz transformation such that the four points $X_i$ (labeled with $i∈${1,2,3,4}) have the same time coordinate in the transformed frame. In the transformed frame, the four points form a tetrahedron in space at some instant in time.

Assuming the tetrahedron has finite volume, find the circumcenter and circumradius. The circumcenter provides the spatial coordinates of the intersection point in the new frame, and the circumradius provides the time coordinate (via time of flight).

Transform back to obtain the coordinates of the intersection point in the original frame.

The computed point $X_c$ should satisfy the four constraints $dX_I^2:=η_{μν}(X_c^μ-X^μ_I)(X_c^ν-X^ν_I)=0$, where $μ$, $ν$ are spacetime indices and $η_{μν}$ is the Minkowski metric.

The case of a timelike hyperplane is conceptually similar, but more intricate:

Find and perform a Lorentz transformation such that the four points $X_I$ (labeled with $I∈$ {1,2,3,4}) lie on a plane $P_z$ specified by a $z$ coordinate in the transformed frame. In the transformed frame, the four points lie on an elliptic hyperboloid.

Find the vertex of the hyperboloid, and the Minkowski distance $R$ from the vertex to a point on the hyperboloid. There are two intersection points. The $z$ coordinate for the intersection points is a distance $R$ in the direction normal to the $P_z$ plane.

Transform back to obtain the coordinates of the intersection point in the original frame.

There are two intersection points in the case of a timelike hyperplane; this is known as the bifurcation problem (Coll et al., Phys. Rev. D 86, 084036 (2012)). The addition of a fifth emission point will in most cases permit the selection of a single emission point.

Multi locator function

cereal.mlocator — Function

Multiple emission set locator

mlocator( X::RealMtx , locator::Function , Nbase::Int , dual::Bool, q::Real )

This computes the intersection point $X_c$ for a large number of emission points. Given an $m×n$ matrix X of $n$ emission points, and a function locator designed to work with Nbase$<n$ emission points, this function is designed to minimize errors and solve the bifurcation problem when needed.

For locator functions designed to work with $N = 4$ emission points, the bifurcation problem is solved by way of a rudimentary clustering algorithm. The clustering algorithm sorts points according to their norms and identifies closely clustered points according to the differences between the neighbors of the sorted points, and the tolerance parameter q.

In all cases, the errors are minimized by sorting points according to their Minkowski norms and selecting the point with the smallest norm.